tests/test_base_tts.py
import pytest
import numpy as np
from src.opentts_tensorflow_integration.core import BaseTTS
from src.opentts_tensorflow_integration.utils import TTSError

class DummyTTS(BaseTTS):
    """A dummy implementation of BaseTTS for testing purposes."""

    def synthesize(self, text: str) -> np.ndarray:
        return np.array([0.0, 0.1, 0.2])  # Dummy audio data

    def load_model(self, model_path: str):
        pass

    @property
    def available_voices(self) -> list:
        return ["voice1", "voice2"]

    def set_voice(self, voice_name: str):
        self.current_voice = voice_name

def test_base_tts_initialization():
    tts = DummyTTS()
    assert tts.sample_rate == 22050
    assert tts.current_voice is None

def test_normalize_text():
    tts = DummyTTS()
    assert tts.normalize_text("Hello, World!") == "hello, world!"
    assert tts.normalize_text("  TESTING  ") == "testing"

def test_normalize_text_error():
    tts = DummyTTS()
    with pytest.raises(TTSError):
        tts.normalize_text(None)

def test_save_audio(tmp_path):
    tts = DummyTTS()
    audio = np.array([0.0, 0.1, 0.2])
    file_path = tmp_path / "test_audio.wav"
    tts.save_audio(audio, str(file_path))
    assert file_path.exists()

def test_save_audio_error(tmp_path):
    tts = DummyTTS()
    audio = np.array([0.0, 0.1, 0.2])
    file_path = tmp_path / "non_existent_dir" / "test_audio.wav"
    with pytest.raises(TTSError):
        tts.save_audio(audio, str(file_path))

def test_abstract_methods():
    with pytest.raises(TypeError):
        BaseTTS()

def test_dummy_tts_implementation():
    tts = DummyTTS()
    assert isinstance(tts.synthesize("test"), np.ndarray)
    assert tts.available_voices == ["voice1", "voice2"]
    tts.set_voice("voice1")
    assert tts.current_voice == "voice1"

# Add more tests as needed